+++ Assets/Script/Helpers/PracticeSectionHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bad4721b90bfdcd4395e140e2116107b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
